feat: pelanggan admin dashboard

+++ app/Http/Controllers/API/UserController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        Gate::authorize('admin', $request->user());
         $user = User::find($id);
         $result = $user->delete();
         if ($result) {
